Programa dividido en ficheros .cpp, .h y ejecutable
