+++ laser_go/package.xml
@@ -55,7 +55,6 @@
   <build_depend>sensor_msgs</build_depend>
   <build_depend>kobuki_msgs</build_depend>
   <build_depend>roslint</build_depend>
-  <build_depend>visualization_msgs</build_depend>
 
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>rospy</build_export_depend>
@@ -63,7 +62,6 @@
   <build_export_depend>sensor_msgs</build_export_depend>
   <build_export_depend>kobuki_msgs</build_export_depend>
   <build_export_depend>roslint</build_export_depend>
-  <build_export_depend>visualization_msgs</build_export_depend>
 
   <exec_depend>roscpp</exec_depend>
   <exec_depend>rospy</exec_depend>
@@ -71,7 +69,6 @@
   <exec_depend>sensor_msgs</exec_depend>
   <exec_depend>kobuki_msgs</exec_depend>
   <exec_depend>roslint</exec_depend>
-  <exec_depend>visualization_msgs</exec_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
